refactor: break out data specific functions into different files
This has some nasty circular dependencies atm
